Applying changes from #1213 to new PvModel
